requirements.txt
streamlit>=1.20.0
PyPDF2>=3.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-anthropic>=0.0.10
langchain-openai>=0.0.5
spacy>=3.5.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.3.0
serpapi>=0.1.5
google-search-results>=2.4.2
sentence_transformers>=3.4.1
gensim>=4.3.3
pinecone>=5.4.2
pinecone-client>=5.0.1
pinecone-plugin-interface>=0.0.7